+++ packages/search-indexer/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.10.0 (2026-08-17)
+
+### 🧱 Updated Dependencies
+
+- Updated @lde/search-typesense to 0.25.0
+- Updated @lde/search-pipeline to 0.22.0
+- Updated @lde/search to 0.22.0
+
 ## 0.9.1 (2026-08-14)
 
 ### 🧱 Updated Dependencies